Refactor user test execution to happen in a background service.
User initialization now can be delayed after a user is "added"

+++ src/Zoxive.HttpLoadTesting.Client/Framework/Core/Schedules/AddUsers.cs
@@ -5,14 +5,14 @@ namespace Zoxive.HttpLoadTesting.Framework.Core.Schedules
 {
     public class AddUsers : ISchedule
     {
-        public AddUsers(int totalUsers, int usersEvery, int seconds)
+        public AddUsers(int totalUsers, int usersEvery, double seconds)
         {
             Users = totalUsers;
             UsersEvery = usersEvery;
             AddUsersEverySeconds = seconds;
         }
 
-        public int AddUsersEverySeconds { get; }
+        public double AddUsersEverySeconds { get; }
 
         public int UsersEvery { get; }
